Mise en place de la redimensionnement des images

+++ src/Framework/Upload.php
@@ -3,13 +3,13 @@
 
 namespace App\Framework;
 
-
+use Intervention\Image\ImageManager;
 use Psr\Http\Message\UploadedFileInterface;
 
 class Upload
 {
     protected $path;
-    protected $formats;
+    protected $formats = [];
 
     public function __construct(?string $path = null)
     {
@@ -20,38 +20,66 @@ class Upload
 
     public function upload(UploadedFileInterface $file, ?string $oldfile = null)
     {
-        $this->delete($oldfile);
-        $filename = $file->getClientFilename();
-        $targetPath = $this->addSuffix($this->path . DIRECTORY_SEPARATOR . $filename);
-        $dirname = pathinfo($targetPath, PATHINFO_DIRNAME);
-        if (!file_exists($dirname)) {
-            mkdir($dirname, 777, true);
+        if ($file->getError() === UPLOAD_ERR_OK) {
+            $this->delete($oldfile);
+            $filename = $file->getClientFilename();
+            $targetPath = $this->addCopySuffix($this->path . DIRECTORY_SEPARATOR . $filename);
+            $dirname = pathinfo($targetPath, PATHINFO_DIRNAME);
+            if (!file_exists($dirname)) {
+                mkdir($dirname, 777, true);
+            }
+            $file->moveTo($targetPath);
+            $this->generateFormats($targetPath);
+            return pathinfo($targetPath)['basename'];
         }
-        $file->moveTo($targetPath);
-        return pathinfo($targetPath)['basename'];
+        return null;
     }
 
-    private function addSuffix(string $targetPath)
+    private function addCopySuffix(string $targetPath)
     {
         if (file_exists($targetPath)) {
-            $info = pathinfo($targetPath);
-            $targetPath = $info['dirname'] .
-                DIRECTORY_SEPARATOR .
-                $info['filename']
-                . '_copy.' .
-                $info['extension'];
-            return $this->addSuffix($targetPath);
+            $targetPath = $this->getPathWithSuffix($targetPath, "copy");
+            return $this->addCopySuffix($targetPath);
         }
         return $targetPath;
     }
 
-    private function delete(?string $oldfile = null)
+    public function delete(?string $oldfile = null)
     {
         if ($oldfile) {
             $oldfile = $this->path . DIRECTORY_SEPARATOR . $oldfile;
             if (file_exists($oldfile)) {
                 unlink($oldfile);
             }
+            foreach ($this->formats as $format => $size) {
+                $oldfileFormat = $this->getPathWithSuffix($oldfile, $format);
+                if (file_exists($oldfileFormat)) {
+                    unlink($oldfileFormat);
+                }
+            }
+        }
+    }
+
+    public function getPathWithSuffix(string $path, string $suffix)
+    {
+        $info = pathinfo($path);
+
+        return $info['dirname'] .
+            DIRECTORY_SEPARATOR .
+            $info['filename']
+            . "_$suffix." .
+            $info['extension'];
+    }
+
+    private function generateFormats(string $targetPath)
+    {
+        foreach ($this->formats as $format => $size) {
+            $destination = $this->getPathWithSuffix($targetPath, $format);
+            $manager = new  ImageManager(['driver' => 'gd']);
+            [$width, $height] = $size;
+            $manager->make($targetPath)
+                ->fit($width, $height)
+                ->save($destination);
         }
     }
 }
